drivers: sensor: opt3001: Add support for trigger
This commit adds support for sensor triggers in the OPT3001 driver.

To utilize this feature, the sensor's INT pin must be connected to the MCU
and configured in the DTS using the newly added `int-gpios` property.

The upper and lower thresholds of the trigger can be set using the
`sensor_attr_set` API.

Signed-off-by: Chen Xingyu <hi@xingrz.me>